feat: add some code example

+++ Array of CPP/1DArray.cpp
@@ -7,8 +7,7 @@ int main() {
     string animals[5] = { "Cats", "Dogs", 
         "Rabbits", "Gerbils", "Hamsters" };
 
-    // Tampilkan index dan nilai tiap elemen
-    for (size_t i = 0; i < 5; ++i) {
+    for (int i = 0; i < 5; ++i) {
         // animals[i] buat panggil array
         cout << "Index " << i << ": " << animals[i] << "\n"; 
     }
